fix: 신뢰도 정확도 오류 수정

See merge request s11-fintech-finance-sub2/S11P22B201!179

+++ backend/stockolm/src/main/java/com/example/stockolm/domain/follow/dto/response/FollowAnalystResponse.java
@@ -8,13 +8,13 @@ public class FollowAnalystResponse {
     private String userName;
     private String userNickName;
     private String userImagePath;
-    private Double reliability;
-    private Double accuracy;
+    private Long reliability;
+    private Long accuracy;
     private Integer totalAnalystRanking;
 
     @Builder
     public FollowAnalystResponse(String userName, String userNickName, String userImagePath,
-                                 Double reliability, Double accuracy, Integer totalAnalystRanking) {
+                                 Long reliability, Long accuracy, Integer totalAnalystRanking) {
         this.userName = userName;
         this.userNickName = userNickName;
         this.userImagePath = userImagePath;
